refactor(pillars-types): single contract-lookup enforcer; split derived-obligation creators; drop dead code
getContract (syntheticContract wins over catalog entry) existed as three
verbatim copies plus four inlined expressions; the rule now lives in one
module imported at all seven sites. [LAW:single-enforcer]

createDerivedObligations fused three unrelated creators — adapter,
cardinality-adapter, and payload-anchor obligations — into one function.
Each is now its own single-purpose creator called distinctly by the
fixpoint driver. [LAW:decomposition]

Dead code removed: apply-elaboration's discharged-id precomputation (never
consumed, and its block-id list wrongly mixed in edge ids) and fixpoint's
empty discharged-obligation loop (plan diagnostics are already collected at
apply time, the single correct point). [LAW:one-source-of-truth]
