+++ src/pillars/types/solve/obligations/index.ts
@@ -1,4 +1,6 @@
 export { createMissingInputObligations } from './create-missing-input-obligations';
-export { createDerivedObligations } from './create-derived-obligations';
+export { createAdapterObligations } from './create-adapter-obligations';
+export { createCardinalityAdapterObligations } from './create-cardinality-adapter-obligations';
+export { createPayloadAnchorObligations } from './create-payload-anchor-obligations';
 export { createCardinalityObligations } from './create-cardinality-obligations';
 export { createCycleBreakObligations } from './create-cycle-break-obligations';
